+++ tests/test_arquivo.py
@@ -4,7 +4,7 @@ from twitter.twitter import TwitterUser
 class TestArquivoCSV(unittest.TestCase):
 	def test_arquivo_existe(self):
 		try:
-			file = open('twitter/autores.csv',newline='')
+			file = open('../twitter/autores.csv',newline='')
 			file.close()
 			arquivo = True
 		except:
